backend/requirements.txt
asgiref==3.6.0
black==23.3.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cryptography==3.4.8
Django==4.2
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
keyring==23.5.0
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg2-binary==2.9.6
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2023.3
PyYAML==5.4.1
requests==2.29.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==1.26.15
zipp==1.0.0
